Update for new Manifest version
