--- src/main/java/com/patchworkmc/mapping/remapper/exception/FatalRemappingException.java
@@ -1,7 +1,0 @@
-package com.patchworkmc.mapping.remapper.exception;
-
-public class FatalRemappingException extends RuntimeException {
-	public FatalRemappingException(String message, Exception exception) {
-		super(message, exception);
-	}
-}
