--- src/main/java/com/patchworkmc/mapping/remapper/exception/MissingMappingException.java
@@ -1,7 +1,0 @@
-package com.patchworkmc.mapping.remapper.exception;
-
-public class MissingMappingException extends Exception {
-	public MissingMappingException(String message) {
-		super(message);
-	}
-}
