+++ src/main/java/com/patchworkmc/patch/StringConstantRemapper.java
@@ -6,7 +6,7 @@ import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 
 import com.patchworkmc.Patchwork;
-import com.patchworkmc.mapping.remapper.exception.AmbiguousMappingException;
+import com.patchworkmc.mapping.remapper.AmbiguousMappingException;
 import com.patchworkmc.mapping.remapper.PatchworkRemapper;
 
 /**
@@ -51,7 +51,7 @@ public class StringConstantRemapper extends ClassVisitor {
 				return name;
 			}
 		} else {
-			name = remapper.getClass(name);
+			name = remapper.getNaiveRemapper().getClass(name);
 		}
 
 		return name;
